Move Sprocket cleanup to app generator
