+++ decidim-generators/lib/decidim/generators/install_generator.rb
@@ -34,22 +34,6 @@ module Decidim
                                default: false,
                                desc: "Add the necessary gems to profile the app"
 
-      def remove_sprockets_requirement
-        gsub_file "config/application.rb", %r{require ['"]rails/all['"]\R}, <<~RUBY
-          require "decidim/rails"
-
-          # Add the frameworks used by your app that are not loaded by Decidim.
-          # require "action_mailbox/engine"
-          # require "action_text/engine"
-          require "action_cable/engine"
-          require "rails/test_unit/railtie"
-        RUBY
-
-        gsub_file "config/environments/development.rb", /config\.assets.*$/, ""
-        gsub_file "config/environments/test.rb", /config\.assets.*$/, ""
-        gsub_file "config/environments/production.rb", /config\.assets.*$/, ""
-      end
-
       def copy_migrations
         rails "decidim:choose_target_plugins", "railties:install:migrations"
         recreate_db if options[:recreate_db]
